+++ ABETAccreditationStation/src/main/java/com/maccari/abet/web/controller/ProgramController.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -15,6 +16,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import com.maccari.abet.domain.entity.ProgramData;
 import com.maccari.abet.domain.entity.web.WebProgram;
 import com.maccari.abet.domain.service.ProgramService;
+import com.maccari.abet.exception.IllegalDownloadException;
 
 @Controller
 @RequestMapping("/program")
@@ -60,20 +62,35 @@ public class ProgramController {
 
 	@RequestMapping(value = "/create", method = RequestMethod.POST, params = "remove")
 	public String removeProgram(@RequestParam(value = "remove", required = true) 
-		int id, Model model) {
+			int id, @RequestHeader(value="referer", defaultValue="") String referer, 
+			Model model) {
+		if(referer == null || referer.isEmpty()) {
+        	return "redirect:/error/";
+        }
+		
 		programService.remove(programService.getById(id));
 
 		return "redirect:/program/index";
 	}
 
 	@RequestMapping(value = "/create", method = RequestMethod.POST, params = "outcome")
-	public String viewOutcomes(@RequestParam(value = "outcome", required = true) int id) {
+	public String viewOutcomes(@RequestParam(value = "outcome", required = true) int id,
+			@RequestHeader(value="referer", defaultValue="") String referer) {
+		if(referer == null || referer.isEmpty()) {
+        	return "redirect:/error/";
+        }
+        
 		return "redirect:/outcome/index?id=" + id;
 	}
 	
 	@RequestMapping(value = "/create", method = RequestMethod.POST, params = "deactivate")
 	public String deactivateProgram(@RequestParam(value = "deactivate", required = true) 
-		int id, Model model) {
+			int id, @RequestHeader(value="referer", defaultValue="") String referer, 
+			Model model) {
+		if(referer == null || referer.isEmpty()) {
+        	return "redirect:/error/";
+        }
+		
 		ProgramData program = programService.getById(id);
 		program.setActive(false);
 		programService.update(program);
@@ -83,7 +100,12 @@ public class ProgramController {
 
 	@RequestMapping(value = "/create", method = RequestMethod.POST, params = "reactivate")
 	public String reactivateProgram(@RequestParam(value = "reactivate", required = true) 
-		int id, Model model) {
+			int id, @RequestHeader(value="referer", defaultValue="") String referer,
+			Model model) {
+		if(referer == null || referer.isEmpty()) {
+        	return "redirect:/error/";
+        }
+		
 		ProgramData program = programService.getById(id);
 		program.setActive(true);
 		programService.update(program);
